cleanup, buildfix
