+++ app/@modal/info/page.tsx
@@ -1,33 +1,11 @@
 import * as React from "react"
-import {
-    Dialog,
-    DialogContent,
-    DialogDescription,
-    DialogHeader,
-    DialogTitle,
-    DialogTrigger,
-  } from "@/components/ui/dialog"
-import { useRouter } from "next/navigation"
 import type { Metadata } from 'next'
- 
+import InfoDialog from "./info-dialog"
+
 export const metadata: Metadata = {
   title: 'OPF | Info',
 }
- 
+
 export default function Info() {
-    const router = useRouter();
-    return (
-        <Dialog open onOpenChange={() => router.back()}>
-            <DialogTrigger>INFO</DialogTrigger>
-            <DialogContent>
-                <DialogHeader>
-                    <DialogTitle>ABOUT OPF</DialogTitle>
-                    <DialogDescription>
-                        [TODO: FILL FROM SANITY IO]
-                    </DialogDescription>
-                    Rerum perferendis at consequatur. Nam et sed velit iure rerum. Nisi ipsam aut dolores rerum impedit ut. Perspiciatis sapiente tempora qui id autem. Quia non error quo in vel. Aut iste consequuntur occaecati iusto perferendis quod provident.
-                </DialogHeader>
-            </DialogContent>
-        </Dialog>
-    )
+    return <InfoDialog />
 }
